Migrate from checks to commit statuses
